refactor: :recycle: Add DialogBook and ListBook components to Book component folder
